refactor: move service factories from api/ to services/ layer (#681)
