+++ src/wikimind/api/routes/admin.py
@@ -7,8 +7,8 @@ from typing import TYPE_CHECKING
 from fastapi import APIRouter, Depends
 
 from wikimind.api.deps import require_admin
-from wikimind.api.services import get_admin_service
 from wikimind.database import get_session
+from wikimind.services.factories import get_admin_service
 
 if TYPE_CHECKING:
     from sqlmodel.ext.asyncio.session import AsyncSession
