+++ src/wikimind/api/routes/capture.py
@@ -4,7 +4,6 @@ from fastapi import APIRouter, Depends, Query
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from wikimind.api.deps import get_current_user_id
-from wikimind.api.services import get_capture_service, get_rss_service
 from wikimind.database import get_session
 from wikimind.models import (
     CaptureKind,
@@ -21,6 +20,7 @@ from wikimind.models import (
     RssPollResponse,
 )
 from wikimind.services.capture import CaptureService
+from wikimind.services.factories import get_capture_service, get_rss_service
 from wikimind.services.rss import RssService
 
 router = APIRouter()
