+++ src/wikimind/api/routes/drafts.py
@@ -12,7 +12,6 @@ from typing import TYPE_CHECKING
 from fastapi import APIRouter, Depends
 
 from wikimind.api.deps import get_current_user_id
-from wikimind.api.services import get_draft_service
 from wikimind.database import get_session
 from wikimind.models import (
     ApproveDraftRequest,
@@ -20,6 +19,7 @@ from wikimind.models import (
     CompilationDraftResponse,
     RejectDraftResponse,
 )
+from wikimind.services.factories import get_draft_service
 
 if TYPE_CHECKING:
     from sqlmodel.ext.asyncio.session import AsyncSession
