+++ src/wikimind/api/routes/lint.py
@@ -7,7 +7,6 @@ from typing import TYPE_CHECKING
 from fastapi import APIRouter, Depends, Query
 
 from wikimind.api.deps import get_current_user_id
-from wikimind.api.services import get_linter_service
 from wikimind.database import get_session
 from wikimind.models import (
     DismissFindingResponse,
@@ -16,6 +15,7 @@ from wikimind.models import (
     LintReportDetail,
     LintRunResponse,
 )
+from wikimind.services.factories import get_linter_service
 
 if TYPE_CHECKING:
     from sqlmodel.ext.asyncio.session import AsyncSession
