+++ src/wikimind/api/routes/tags.py
@@ -5,12 +5,12 @@ from fastapi import APIRouter, Depends
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from wikimind.api.deps import get_current_user_id
-from wikimind.api.services import get_tag_service, get_wiki_service
 from wikimind.database import get_session
 from wikimind.models import (
     CreateTagRequest,
     TagResponse,
 )
+from wikimind.services.factories import get_tag_service, get_wiki_service
 from wikimind.services.tags import TagService
 from wikimind.services.wiki import WikiService
