+++ src/wikimind/api/routes/wiki.py
@@ -5,13 +5,6 @@ from fastapi import APIRouter, Depends, Query
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from wikimind.api.deps import get_current_user_id
-from wikimind.api.services import (
-    get_contradiction_service,
-    get_linter_service,
-    get_search_service,
-    get_tag_service,
-    get_wiki_service,
-)
 from wikimind.database import get_session
 from wikimind.models import (
     ArticleEditRequest,
@@ -42,6 +35,13 @@ from wikimind.models import (
     WikilinkMatch,
 )
 from wikimind.services.contradiction import ContradictionService
+from wikimind.services.factories import (
+    get_contradiction_service,
+    get_linter_service,
+    get_search_service,
+    get_tag_service,
+    get_wiki_service,
+)
 from wikimind.services.linter import LinterService
 from wikimind.services.search import SearchService
 from wikimind.services.tags import TagService
