+++ tests/unit/test_compilation_schema_service.py
@@ -7,9 +7,9 @@ from typing import TYPE_CHECKING
 import pytest
 
 from tests.conftest import TEST_USER_ID
-from wikimind.api.services import get_compilation_schema_service
 from wikimind.errors import NotFoundError
 from wikimind.services.compilation_schema import CompilationSchemaService
+from wikimind.services.factories import get_compilation_schema_service
 
 if TYPE_CHECKING:
     from sqlmodel.ext.asyncio.session import AsyncSession
