+++ tests/unit/test_search.py
@@ -8,8 +8,8 @@ from sqlmodel import SQLModel
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from tests.conftest import TEST_USER_ID
-from wikimind.api.services import get_search_service
 from wikimind.models import Article, SearchResponse, SearchResult, User
+from wikimind.services.factories import get_search_service
 from wikimind.services.search import (
     SearchService,
     _article_id_to_rowid,
