+++ tests/unit/test_view_original.py
@@ -9,10 +9,10 @@ from httpx import ASGITransport, AsyncClient
 
 from tests.conftest import TEST_USER_ID
 from wikimind.api.deps import get_current_user_id
-from wikimind.api.services import get_ingest_service
 from wikimind.database import get_session
 from wikimind.main import app
 from wikimind.models import Source, SourceType
+from wikimind.services.factories import get_ingest_service
 from wikimind.storage import LocalFileStorage
